splut rust code; added rust unit test
